feat: refactor editor's collaboration to reduce package size and minimize errors
